+++ src/components/InstallationPage/InstallationPage.scss
