+++ src/components/InstallationPage/components/IntegrationScheme/IntegrationScheme.scss
