+++ src/components/InstallationPage/components/IntegrationScheme/SchemeHeader.jsx
