+++ src/components/InstallationPage/components/IntegrationScheme/TemporaryFile.scss
