+++ src/components/InstallationPage/components/IntegrationScheme/httpInfo.js
