+++ src/components/InstallationPage/components/IntegrationScheme/index.js
