+++ src/components/InstallationPage/components/IntegrationScheme/schemeData.js
