+++ src/components/InstallationPage/components/Notice/Notice.tsx
