+++ src/components/InstallationPage/components/Notice/index.js
