--- src/components/installation/index.js
@@ -1,1 +1,0 @@
-export * from './Installation';
